commit corrigir passagem do id do projeeto gerar nome do arquivo.

+++ src/main/java/br/gov/es/siscap/service/ProspeccaoService.java
@@ -118,7 +118,7 @@ public class ProspeccaoService {
 
 		List<String> emailsInteressadosList = prospeccaoInteressadoService.buscarEmailsInteressadosPorPropeccao(this.buscar(id));
 
-		String nomeArquivo = this.projetoService.gerarNomeArquivo(id) + ".pdf";
+		String nomeArquivo = this.projetoService.gerarNomeArquivo(id.intValue()) + ".pdf";
 
 		boolean confirmacaoEnvioEmail = emailService.enviarEmail(prospeccaoDetalhesDto, emailsInteressadosList, nomeArquivo);
